Memperketat Role pengguna

+++ attr/Admin/report/deleted.php
@@ -1,5 +1,15 @@
 <?php 
 
+// Cek Dulu Role Anggota 
+if ($row['role'] == 'Anggota') {
+    echo "<script>
+    document.location.href = '".base_url."/attr/User'
+  </script>";
+  exit();
+
+  return false;
+
+}
 
 if (deleteReport($_POST) > 0) {
     echo showMessage('success','Berhasil Dihapus','Report Telah Dihapus');
